+++ src/chunk.cpp
@@ -8,7 +8,8 @@ using namespace MineyCrafty;
 
 Chunk::Chunk() : Chunk(0, 0) { }
 
-Chunk::Chunk(int x, int y) : chunkX(x), chunkY(y) {
+Chunk::Chunk(int x, int y) : chunkX(x), chunkY(y)
+{
     xPos = chunkX * Chunk::w;
     yPos = chunkY * Chunk::l;
 }
@@ -16,7 +17,8 @@ Chunk::Chunk(int x, int y) : chunkX(x), chunkY(y) {
 Chunk::~Chunk() { }
 
 void Chunk::init() {
-    if (shader == nullptr && atlas == nullptr) {
+    if (shader == nullptr && atlas == nullptr)
+    {
 		shader = make_unique<Shader>(
 			util::joinPath({ "resources", "vert.glsl" }),
 			util::joinPath({ "resources", "frag.glsl" }),
@@ -37,7 +39,8 @@ void Chunk::init() {
 }
 
 void Chunk::update(std::shared_ptr<Chunk> left, std::shared_ptr<Chunk> right, 
-                  std::shared_ptr<Chunk> forward, std::shared_ptr<Chunk> backward) {
+                  std::shared_ptr<Chunk> forward, std::shared_ptr<Chunk> backward)
+{
     vector<GLfloat> coords = atlas->generateTexCoords(
         pair{0, 0}, pair{1, 0}, pair{1, 0}, pair{1, 0}, pair{1, 0}, pair{2, 0}
     );
@@ -49,33 +52,40 @@ void Chunk::update(std::shared_ptr<Chunk> left, std::shared_ptr<Chunk> right,
     vector<glm::vec2> texcoords;
     float chunkModifierX = static_cast<float>(chunkX * Chunk::w * 2);
     float chunkModifierY = static_cast<float>(chunkY * Chunk::l * 2);
-    for (const auto& pos : blocks) {
+    for (const auto& pos : blocks)
+    {
 		auto&[x, y, z] = pos;
 
         bool blockAbove = y + 1 < Chunk::h && bools[z][y + 1][x];
         bool blockBelow = y == 0  ? true : bools[z][y - 1][x];
 
         bool blockRight = x + 1 < Chunk::w ? bools[z][y][x + 1] : true;
-        if (x + 1 == Chunk::w && right != nullptr) {
+        if (x + 1 == Chunk::w && right != nullptr)
+        {
             blockRight = right->isBlockAt(0, y, z);
         }
         bool blockLeft = x == 0 ? true : bools[z][y][x - 1];
-        if (x == 0 && left != nullptr) {
+        if (x == 0 && left != nullptr)
+        {
             blockLeft = left->isBlockAt(Chunk::w - 1, y, z);
         }
         bool blockInfront = z + 1 < Chunk::l ? bools[z + 1][y][x] : true;
-        if (z + 1 == Chunk::l && forward != nullptr) {
+        if (z + 1 == Chunk::l && forward != nullptr)
+        {
             blockInfront = forward->isBlockAt(x, y, 0);
         }
         bool blockBehind = z == 0 ? true : bools[z - 1][y][x];
-        if (z == 0 && backward != nullptr) {
+        if (z == 0 && backward != nullptr)
+        {
             blockBehind = backward->isBlockAt(x, y, Chunk::l - 1);
         }
         
 		
 		if (!(blockLeft && blockRight && blockInfront && blockBehind
-			&& blockAbove && blockBelow)) {
-			for (int vert = 0; vert < sizeof(cubeVertices) / sizeof(GLfloat); vert += 3) {
+			&& blockAbove && blockBelow))
+        {
+			for (int vert = 0; vert < sizeof(cubeVertices) / sizeof(GLfloat); vert += 3)
+            {
 				vertices.push_back(glm::vec3(
 					cubeVertices[vert] + x * 2 + chunkModifierX,
 					cubeVertices[vert + 1] + y * 2,
@@ -85,13 +95,17 @@ void Chunk::update(std::shared_ptr<Chunk> left, std::shared_ptr<Chunk> right,
 					cubeNormals[vert], cubeNormals[vert + 1], cubeNormals[vert + 2]
 				));
 			}
-			if (blockAbove) {
-				for (size_t c = 0; c < coordsNoGrass.size(); c += 2) {
+			if (blockAbove)
+            {
+				for (size_t c = 0; c < coordsNoGrass.size(); c += 2)
+                {
 					texcoords.push_back(glm::vec2(coordsNoGrass[c], coordsNoGrass[c + 1]));
 				}
 			}
-			else {
-				for (size_t c = 0; c < coords.size(); c += 2) {
+			else
+            {
+				for (size_t c = 0; c < coords.size(); c += 2)
+                {
 					texcoords.push_back(glm::vec2(coords[c], coords[c + 1]));
 				}
 			}
@@ -108,14 +122,16 @@ void Chunk::update(std::shared_ptr<Chunk> left, std::shared_ptr<Chunk> right,
     glBufferData(GL_ARRAY_BUFFER, texcoords.size() * sizeof(glm::vec2), &texcoords[0], GL_STATIC_DRAW);
 	glBindBuffer(GL_ARRAY_BUFFER, normal);
 	glBufferData(GL_ARRAY_BUFFER, normals.size() * sizeof(glm::vec3), &normals[0], GL_STATIC_DRAW);
-	num_verts = vertices.size();
+	num_verts = static_cast<int>(vertices.size());
 }
 
-void Chunk::generateData() {
+void Chunk::generateData()
+{
     blockBool();
 }
 
-void Chunk::draw(glm::mat4& trans, glm::mat4& lightSpace, GLuint depthMap) {
+void Chunk::draw(glm::mat4& trans, glm::mat4& lightSpace, GLuint depthMap)
+{
     auto cam = Camera::getMainCamera();
     shader->activate(trans, cam->getView(), cam->getProj());
 	shader->setMat4("lightSpace", lightSpace);
@@ -143,7 +159,8 @@ void Chunk::draw(glm::mat4& trans, glm::mat4& lightSpace, GLuint depthMap) {
 	shader->deactivate();
 }
 
-void Chunk::drawShadow(glm::mat4& trans, glm::mat4& lightSpace) {
+void Chunk::drawShadow(glm::mat4& trans, glm::mat4& lightSpace)
+{
 	auto cam = Camera::getMainCamera();
 	shadowShader->activate(trans, cam->getView(), cam->getProj());
 	shadowShader->setMat4("lightSpace", lightSpace);
@@ -157,41 +174,50 @@ void Chunk::drawShadow(glm::mat4& trans, glm::mat4& lightSpace) {
 	shadowShader->deactivate();
 }
 
-void Chunk::addCube(int x, int y, int z) {
+void Chunk::addCube(int x, int y, int z)
+{
     blocks.push_back(make_tuple(x, y, z));
 }
 
-void Chunk::finalise() {
+void Chunk::finalise()
+{
     glDeleteBuffers(1, &vbo);
     glDeleteBuffers(1, &texcoord);
     glDeleteVertexArrays(1, &vao);
 }
 
-pair<int, int> Chunk::getPos() {
+pair<int, int> Chunk::getPos()
+{
     return pair{chunkX, chunkY};
 }
 
-void Chunk::blockBool() {
+void Chunk::blockBool()
+{
     bools.reserve(Chunk::h);
-    for (int z = 0; z < Chunk::w; ++z) {
+    for (int z = 0; z < Chunk::w; ++z)
+    {
         vector<vector<bool>> row;
         row.reserve(Chunk::w);
-        for (int y = 0; y < Chunk::h; ++y) {
+        for (int y = 0; y < Chunk::h; ++y)
+        {
             vector<bool> col;
             col.reserve(Chunk::h);
-            for (int x = 0; x < Chunk::l; ++x) {
+            for (int x = 0; x < Chunk::l; ++x)
+            {
                 col.push_back(false);
             }
             row.push_back(col);
         }
         bools.push_back(row);
     }
-    for (const auto& pos : blocks) {
+    for (const auto& pos : blocks)
+    {
         const auto& [x, y, z] = pos;
         bools[z][y][x] = true;
     }
 }
 
-bool Chunk::isBlockAt(int x, int y, int z) {
+bool Chunk::isBlockAt(int x, int y, int z)
+{
     return bools[z][y][x];
 }
